better rerun handling, more logging
if an episode file already exists, stat it and head the url, so we can
skip epiosodes we've already completely downloaded.  we apparently can't
trust the enclosure's stated length.

if we hit an error in the copy-from-shapeio-reader-to-file operation, we
remove whatever we downloaded, under the assumption what we did download
was mangled.